aug methods moved to tranforms
